refactor(pkg/common): remove common package in favor of specialized packages
moved common.config to configger package, and common.logger to logger package
