+++ pkg/logger/logger.go
@@ -1,20 +1,21 @@
-package common
+package logger
 
 import (
 	"sync"
 
+	"github.com/joao-fontenele/go-url-shortener/pkg/configger"
 	"go.uber.org/zap"
 )
 
 var logger *zap.Logger
 var once sync.Once
 
-// GetLogger gets singleton logger instance
-func GetLogger() *zap.Logger {
+// Get gets singleton logger instance
+func Get() *zap.Logger {
 	once.Do(func() {
 		var err error
 
-		env := GetConf().Env
+		env := configger.Get().Env
 		if env == "development" {
 			logger, err = zap.NewDevelopment()
 		} else if env == "test" {
